Added drawing capbility
This change enables drawing of lines on a PDF document at a given page
number.

Currently only black line is drawn, but the ability to specify any
colour using RGB scheme should be allowed. Along with specifying the
line thickness of the line to be drawn. The functions that handle that
are already implemented it's just deciding how to pass those arguments
to draw-line function so that it doesn't become too verbose. As in, too
many parameters that it has to take.