Added the namespace 'common'
The idea is that it should hold the functions that are frequently used
by the other namespaces. Such as PDF reading.

In `is-pdf?' function I'm using PreflightParser as it seems to do most
of the heavy lifting already, but in it's private methods, where it
reads the first two top lines of a file and tries to determine if they
comply with the PDF standard. I just went with this approach of
try/catch as I see that there is no need to repeat that work again.