Remove org.apache.pdfbox/preflight dependency
The idea is to remove the external dependency on
org.apache.pdfbox/preflight.

The reason for that is since there is one less dependency, it's one less
thing to worry about when the project needs dependency updates.

And the check this dependency performed is something light so why not
have it on it's own.