Upgraded to PDFBox 2.0.0 and Clojure 1.8.0
Borrowed some code from Markus Penttila <markus.penttila@solita.fi> as
he has already done the upgrade in his fork.

Like Markus, I've upgraded the PDF files used for tests, by taking in
his changes.

On one file the issue was a bug with fonts embedded in the PDF
document. Since this bug was a PDFBox issue:

https://issues.apache.org/jira/browse/PDFBOX-2661

I don't wish to bang head about it for now. So to 'work' around it, I've
changed the PDF file in the tests by renaming the file and keeping it
around for some tests. But for the ones it was failing on, I added a
simpler file without the font issue.

Also for the tests, this message pops up:

  $ lein test pdfboxing.form-test
  Mar 21, 2016 9:50:39 PM org.apache.pdfbox.cos.COSDocument finalize
  WARNING: Warning: You did not close a PDF Document
  ...

It seems to only come up for the tests that test invalid input or an
error/exception. So for now it's OK to ignore them.