Fix for split-pdf
Previously with-open was used to open the PDF document that would be
split. Which is fine if you're just thinking about closing the file
after you've split it. But it doesn't seem very useful, because once
you've split the file, you probably want to do something with it.

I'm reverting this change introduced via
2ed58248360788cbdd4a4b5dbc95eadaddd9c767 just to get some opinions in
this issue: https://github.com/dotemacs/pdfboxing/issues/46