Unifed the jobs in one YAML file
Since they run in two separate file, it produces two actions with the
same name, that of the subject of the commit. Which is confusing, so
joined the jobs in one YAML file because that presents a cleaner
overview in GitHub Actions web UI, as you can see the names of all the
jobs and the outcome.