Updated the CHANGELOG
Lots of changes went in, which I've added to the CHANGELOG. But not
all of them had link backs to the relevant PRs. And some of them where
changes that were added in separate branches, that never went through
the PR workflow.

I went through all the changes and the PRs and added all the relevant
links so that it can be easily referenced.